feat:IGED-138: add delete local image from path in ImageLocalDataSource, add create mission unit tests

+++ LaGrandeEcoleDuDroit/Common/Data/Local/ImageLocalDataSource.swift
@@ -21,7 +21,15 @@ class ImageLocalDataSource {
     
     func deleteLocalImage(folderName: String, fileName: String) async throws {
         if let imageUrl = getFolderUrl(folderName: folderName)?.appendingPathComponent(fileName) {
-            try FileManager.default.removeItem(at: imageUrl)
+            if FileManager.default.fileExists(atPath: imageUrl.path()) {
+                try FileManager.default.removeItem(at: imageUrl)
+            }
+        }
+    }
+    
+    func deleteLocalImage(imagePath: String) async throws {
+        if FileManager.default.fileExists(atPath: imagePath) {
+            try FileManager.default.removeItem(atPath: imagePath)
         }
     }
